Refactoring and bugfix
Bugfix
- removed constructor param with DataCookerException class

Refactoring
- extract db config in test directory
