+++ test/DataStore/ComplexTest.php
@@ -7,11 +7,14 @@ use battlecook\DataStore\Buffer;
 use battlecook\DataStore\RelationDatabase;
 use PHPUnit\Framework\TestCase;
 use test\Fixture\DataStorage\Item;
+use test\Helper\DatabaseConfigTrait;
 
 require __DIR__ . '/../../vendor/autoload.php';
 
 class ComplexTest extends TestCase
 {
+    use DatabaseConfigTrait;
+
     /**
      * @throws \battlecook\DataCookerException
      */
